devex: refactor for coverage
splitting functionality in multiple files to enable better testability and coverage
